rename files with prefix nikss
