+++ CLI/register.c
@@ -21,7 +21,7 @@
 
 #include <jansson.h>
 
-#include <psabpf.h>
+#include <nikss.h>
 
 #include "register.h"
