+++ include/nikss.h
