+++ lib/nikss_direct_counter.c
@@ -20,7 +20,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <psabpf.h>
+#include <nikss.h>
 
 void nikss_direct_counter_ctx_init(nikss_direct_counter_context_t *dc_ctx)
 {
