+++ lib/nikss_meter.c
@@ -24,12 +24,12 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-#include <psabpf.h>
+#include <nikss.h>
 
 #include "btf.h"
 #include "common.h"
-#include "psabpf_meter.h"
-#include "psabpf_table.h"
+#include "nikss_meter.h"
+#include "nikss_table.h"
 
 /**
  * This function comes from DPDK
